Factor out branding utils and config into their own files
