+++ src/components/FooterWrapper.js
@@ -2,8 +2,8 @@ import { Fragment, useEffect, useState } from 'react'
 import { a, div, h } from 'react-hyperscript-helpers'
 import { Clickable, Link } from 'src/components/common'
 import { icon } from 'src/components/icons'
+import { isBioDataCatalyst } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
-import { isBioDataCatalyst } from 'src/libs/config'
 import { footerLogo } from 'src/libs/logos'
 import * as Nav from 'src/libs/nav'
 import * as Utils from 'src/libs/utils'
