+++ src/components/NpsSurvey.js
@@ -7,8 +7,8 @@ import { icon } from 'src/components/icons'
 import { TextArea } from 'src/components/input'
 import Interactive from 'src/components/Interactive'
 import { Ajax } from 'src/libs/ajax'
+import { getEnabledBrand } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
-import { getEnabledBrand } from 'src/libs/config'
 import { withErrorIgnoring } from 'src/libs/error'
 import { useStore } from 'src/libs/react-utils'
 import { authStore, userStatus } from 'src/libs/state'
