+++ src/libs/auth.js
@@ -6,7 +6,8 @@ import { div, h } from 'react-hyperscript-helpers'
 import { FrameworkServiceLink, ShibbolethLink, UnlinkFenceAccount } from 'src/components/common'
 import { cookiesAcceptedKey } from 'src/components/CookieWarning'
 import { Ajax, fetchOk } from 'src/libs/ajax'
-import { getConfig, getEnabledBrand } from 'src/libs/config'
+import { getEnabledBrand } from 'src/libs/brand-utils'
+import { getConfig } from 'src/libs/config'
 import { withErrorIgnoring, withErrorReporting } from 'src/libs/error'
 import { captureAppcuesEvent } from 'src/libs/events'
 import * as Nav from 'src/libs/nav'
