+++ src/pages/LandingPage.js
@@ -5,8 +5,9 @@ import { icon } from 'src/components/icons'
 import covidHero from 'src/images/covid-hero.jpg'
 import hexButton from 'src/images/hex-button.svg'
 import terraHero from 'src/images/terra-hero.png'
+import { getEnabledBrand, isFirecloud, isTerra } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
-import { getEnabledBrand, isDataBrowserFrontPage, isFirecloud, isTerra } from 'src/libs/config'
+import { isDataBrowserFrontPage } from 'src/libs/config'
 import * as Nav from 'src/libs/nav'
 import { setLocalPref } from 'src/libs/prefs'
 import * as Style from 'src/libs/style'
