+++ src/pages/SignIn.js
@@ -2,8 +2,8 @@ import { Fragment } from 'react'
 import { div, h, p } from 'react-hyperscript-helpers'
 import { HeroWrapper, Link } from 'src/components/common'
 import SignInButton from 'src/components/SignInButton'
+import { getEnabledBrand, isAnvil, isBioDataCatalyst, isElwazi, isFirecloud } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
-import { getEnabledBrand, isAnvil, isBioDataCatalyst, isElwazi, isFirecloud } from 'src/libs/config'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
