+++ src/pages/workspaces/workspace/WorkspaceContainer.js
@@ -15,8 +15,9 @@ import { TabBar } from 'src/components/tabBars'
 import TopBar from 'src/components/TopBar'
 import { Ajax, saToken } from 'src/libs/ajax'
 import { getUser } from 'src/libs/auth'
+import { isTerra } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
-import { isAnalysisTabVisible, isTerra } from 'src/libs/config'
+import { isAnalysisTabVisible } from 'src/libs/config'
 import { withErrorIgnoring, withErrorReporting } from 'src/libs/error'
 import * as Nav from 'src/libs/nav'
 import { clearNotification, notify } from 'src/libs/notifications'
